requirements.txt
torch>=2.2
transformers>=4.46
datasets>=2.18
peft>=0.13
accelerate>=0.33
bitsandbytes>=0.43; platform_system != "Windows"
scikit-learn>=1.4
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
pyyaml>=6.0
tqdm>=4.66
